# tests/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelNet",
    "description": "校园网加密流量入侵检测系统",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "system",
          "system"
        ],
        "summary": "获取当前配置",
        "description": "获取当前配置 — 从 Settings 单例读取。",
        "operationId": "get_settings_endpoint_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "system",
          "system"
        ],
        "summary": "更新阈值配置",
        "description": "更新阈值配置 — 运行期覆盖 + DB 持久化 + 检测器热重建。",
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThresholdUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/traffic": {
      "get": {
        "tags": [
          "traffic",
          "traffic"
        ],
        "summary": "获取实时流量统计",
        "description": "获取实时流量统计数据（从 RuntimeState 读取内存中的当前值）。",
        "operationId": "get_traffic_api_traffic_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrafficStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/alerts": {
      "get": {
        "tags": [
          "alerts",
          "alerts"
        ],
        "summary": "查询告警列表（支持分页和级别筛选）",
        "description": "获取告警列表（从 SQLite alerts 表查询）。",
        "operationId": "get_alerts_api_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "级别筛选: high/medium/low/all",
              "default": "all",
              "title": "Level"
            },
            "description": "级别筛选: high/medium/low/all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/stats": {
      "get": {
        "tags": [
          "alerts",
          "alerts"
        ],
        "summary": "获取告警统计",
        "description": "获取告警统计（总数、类型分布、级别分布）。",
        "operationId": "get_alert_stats_api_alerts_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks",
          "tasks"
        ],
        "summary": "获取所有任务状态",
        "description": "获取所有任务状态。\n\n取代旧端点：\n- GET /api/capture/status\n- GET /api/capture/enhanced-status\n- GET /api/detector/status\n- GET /api/attack/status\n- GET /api/auto/status\n- GET /api/dual/stats\n- GET /api/model/train-status",
        "operationId": "list_tasks_api_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks/{name}/start": {
      "post": {
        "tags": [
          "tasks",
          "tasks"
        ],
        "summary": "启动指定任务",
        "description": "启动指定任务。\n\n限时任务可指定 duration（秒），未指定则使用任务默认时长。\n连续任务忽略 duration 参数。\n\n取代旧端点：\n- POST /api/capture/start\n- POST /api/capture/start-enhanced\n- POST /api/detector/start\n- POST /api/attack/start\n- POST /api/auto/start\n- POST /api/demo/start\n- POST /api/dual/load\n- POST /api/model/train",
        "operationId": "start_task_api_tasks__name__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models": {
      "get": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "获取可用模型列表",
        "description": "获取可用模型列表 — 从 registry.json 读取。",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/models/{name}": {
      "delete": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "删除指定模型",
        "description": "删除指定模型 — 从 registry.json 移除并删除 run 目录。\n\n悬挂指针修复（2026-09-19）：删除后按 `model/train.py` 自身规则重算\n`latest.json` / `best.json`，并把剩余条目的 `is_best` 与 best 对齐；\n删空后指针文件一并移除。注册表写失败（500）时不动指针文件。",
        "operationId": "delete_model_api_models__name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios": {
      "get": {
        "tags": [
          "scenarios",
          "scenarios"
        ],
        "summary": "获取可用剧本列表",
        "description": "获取可用剧本列表（供前端抽屉渲染，替代硬编码按钮）。",
        "operationId": "list_scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/scenarios/start": {
      "post": {
        "tags": [
          "scenarios",
          "scenarios"
        ],
        "summary": "启动指定剧本",
        "description": "启动指定剧本（按序执行子任务，任一失败则整体失败并回滚）。",
        "operationId": "start_scenario_api_scenarios_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/scenarios/stop": {
      "post": {
        "tags": [
          "scenarios",
          "scenarios"
        ],
        "summary": "停止指定剧本",
        "description": "停止指定剧本（停止所有子任务，幂等）。",
        "operationId": "stop_scenario_api_scenarios_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioStopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/tls/stats": {
      "get": {
        "tags": [
          "tls",
          "tls"
        ],
        "summary": "TLS加密流量统计",
        "description": "获取 TLS 加密流量统计。",
        "operationId": "get_tls_stats_api_tls_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Tls Stats Api Tls Stats Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tls/suspicious": {
      "get": {
        "tags": [
          "tls",
          "tls"
        ],
        "summary": "获取可疑TLS记录",
        "description": "获取可疑 TLS 记录列表。",
        "operationId": "get_tls_suspicious_api_tls_suspicious_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Tls Suspicious Api Tls Suspicious Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payload/check": {
      "post": {
        "tags": [
          "payload",
          "payload"
        ],
        "summary": "载荷送检（旧路径）",
        "description": "对 HTTP 载荷执行应用层检测（SQL 注入 + XSS）。\n\n保留旧路径 /api/payload/check 以兼容现有客户端。",
        "operationId": "check_payload_api_payload_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayloadAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayloadAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/payload/analyze": {
      "post": {
        "tags": [
          "payload",
          "payload"
        ],
        "summary": "载荷分析",
        "description": "对 HTTP 载荷执行应用层检测（SQL 注入 + XSS）。",
        "operationId": "analyze_payload_api_payload_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayloadAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayloadAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stream": {
      "get": {
        "tags": [
          "stream",
          "stream"
        ],
        "summary": "SSE 实时事件流",
        "description": "单条 SSE 事件流。\n\n通过 ?topics= 参数选择订阅主题。**帧的 `event:` 名与 topic 名相同。**",
        "operationId": "event_stream_api_stream_get",
        "parameters": [
          {
            "name": "topics",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "订阅主题（逗号分隔）",
              "default": "traffic,alert",
              "title": "Topics"
            },
            "description": "订阅主题（逗号分隔）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "会话登录",
        "description": "用户名+密码登录，成功后写入会话。",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SettingsResponse": {
        "properties": {
          "thresholds": {
            "additionalProperties": true,
            "type": "object",
            "title": "Thresholds",
            "description": "规则检测阈值"
          },
          "ml_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ml Config",
            "description": "ML 检测配置"
          },
          "web_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Web Config",
            "description": "Web 面板配置"
          },
          "auth_enabled": {
            "type": "boolean",
            "title": "Auth Enabled",
            "default": false
          },
          "alert_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Alert Config",
            "description": "告警配置"
          }
        },
        "type": "object",
        "title": "SettingsResponse",
        "description": "配置响应 — 对齐 Settings 类字段。",
        "examples": [
          {
            "alert_config": {
              "max_alert_api_return": 20
            },
            "auth_enabled": false,
            "ml_config": {
              "ml_conf_high": 0.7,
              "ml_interval_sec": 5.0
            },
            "thresholds": {
              "ddos_threshold": 500,
              "port_scan_threshold": 50
            },
            "web_config": {
              "web_port": 5000
            }
          }
        ]
      },
...
      "ThresholdUpdateRequest": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "阈值键名"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "阈值新值"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "ThresholdUpdateRequest",
        "description": "阈值更新请求。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}